Add PCM5102 AudioSink, move commonalities into BufferedAudioSink
