+++ targets/esp32/main/CMakeLists.txt
@@ -1,7 +1,9 @@
 set(COMPONENT_SRCS
     "main.cpp"
     "sinks/AC101AudioSink.cpp"
+    "sinks/BufferedAudioSink.cpp"
     "sinks/ES9018AudioSink.cpp"
+    "sinks/PCM5102AudioSink.cpp"
     "sinks/ac101.c"
 )
 set(COMPONENT_ADD_INCLUDEDIRS "." "sinks")
