+++ targets/esp32/main/sinks/PCM5102AudioSink.cpp
@@ -1,8 +1,8 @@
-#include "ES9018AudioSink.h"
+#include "PCM5102AudioSink.h"
 
 #include "driver/i2s.h"
 
-ES9018AudioSink::ES9018AudioSink()
+PCM5102AudioSink::PCM5102AudioSink()
 {
     i2s_config_t i2s_config = {
 
@@ -10,13 +10,12 @@ ES9018AudioSink::ES9018AudioSink()
         .sample_rate = 44100,
         .bits_per_sample = (i2s_bits_per_sample_t)16,
         .channel_format = I2S_CHANNEL_FMT_RIGHT_LEFT, //2-channels
-        .communication_format = (i2s_comm_format_t)I2S_COMM_FORMAT_STAND_MSB,
+        .communication_format = (i2s_comm_format_t)I2S_COMM_FORMAT_I2S,
         .intr_alloc_flags = 0, //Default interrupt priority
         .dma_buf_count = 8,
         .dma_buf_len = 512,
         .use_apll = true,
         .tx_desc_auto_clear = true, //Auto clear tx descriptor on underflow
-        .fixed_mclk = 384 * 44100
     };
 
     i2s_pin_config_t pin_config = {
@@ -27,14 +26,10 @@ ES9018AudioSink::ES9018AudioSink()
     };
     i2s_driver_install((i2s_port_t)0, &i2s_config, 0, NULL);
     i2s_set_pin((i2s_port_t)0, &pin_config);
-    
-    PIN_FUNC_SELECT(PERIPHS_IO_MUX_GPIO0_U, FUNC_GPIO0_CLK_OUT1);
-    REG_SET_FIELD(PIN_CTRL, CLK_OUT1, 0);
-    ESP_LOGI("OI", "MCLK output on CLK_OUT1");
-    
+
     startI2sFeed();
 }
 
-ES9018AudioSink::~ES9018AudioSink()
+PCM5102AudioSink::~PCM5102AudioSink()
 {
 }
